Improved implementation of content.py
